moved split function
